guassian blur tested and functional, added getDerivativeBrightness() method which finds change in brightness at given points which is used to determine when an edge is found in the image, software produces an image showing where it beleives edges are.
Original Grayscale conversion method omitted for an alternative version in hopes of an alternative conversion being better an extenuating contrast along the edge of the face. It's been determined a monotone background is necessary for best results and minimal background edges.

Considering not using black and white, but color to better indetify contrast in images. Maybe instead of using brightness I use contrast in color?