tested use of color vs black and white, not clear if one is better than the other. For testing purposes have been using black and white.
Introduced the drawOutline() method, currently seems to be mostly disfunctional because it is just trying a line perpendicular to a contrast edge. I think it would be more effective to find two points along the same edge and connect them with lines. This may yield better results.

Created a new constructor call with more parameters for

Additionally added more in-method variables as static variables initialized upon a constructor call.

I think adding in a feature to save settings, where when saved they are written to a text file that is then reread would be very useful.

Also need to be thinking about multithreading...